chore: env check func moved to utils file
